requirements.txt
torch>=2.2.0
numpy>=1.26.0
scipy>=1.13.0
scikit-learn>=1.4.0
h5py>=3.11.0
onnx>=1.16.0
onnxruntime>=1.18.0
onnxscript>=0.1.0
pandas>=2.2.0
matplotlib>=3.9.0
pyyaml>=6.0
pytest>=8.0.0
